+++ src/go/rpk/pkg/cli/debug/bundle/BUILD
@@ -21,7 +21,6 @@ go_library(
             "//src/go/rpk/pkg/adminapi",
             "//src/go/rpk/pkg/cli/debug/common",
             "//src/go/rpk/pkg/config",
-            "//src/go/rpk/pkg/httpapi",
             "//src/go/rpk/pkg/kafka",
             "//src/go/rpk/pkg/os",
             "//src/go/rpk/pkg/out",
@@ -86,7 +85,6 @@ go_library(
             "//src/go/rpk/pkg/adminapi",
             "//src/go/rpk/pkg/cli/debug/common",
             "//src/go/rpk/pkg/config",
-            "//src/go/rpk/pkg/httpapi",
             "//src/go/rpk/pkg/kafka",
             "//src/go/rpk/pkg/os",
             "//src/go/rpk/pkg/out",
