+++ src/go/rpk/pkg/cli/debug/remotebundle/download.go
@@ -41,6 +41,7 @@ func newDownloadCommand(fs afero.Fs, p *config.Params) *cobra.Command {
 		noConfirm bool
 		jobID     string
 		outFile   string
+		uploadURL string
 	)
 	cmd := &cobra.Command{
 		Use:   "download",
@@ -103,11 +104,18 @@ Use the flag '--no-confirm' to avoid the confirmation prompt.
 				os.Exit(1)
 			}
 			fmt.Printf("\nSuccessfully downloaded remote debug bundle to %v\n", downloadPath)
+
+			if uploadURL != "" {
+				err = common.UploadBundle(cmd.Context(), downloadPath, uploadURL)
+				out.MaybeDie(err, "unable to upload bundle: %v", err)
+				fmt.Println("Successfully uploaded the bundle")
+			}
 		},
 	}
 	cmd.Flags().StringVarP(&outFile, "output", "o", "", "The file path where the debug file will be written (default ./<timestamp>-remote-bundle.zip)")
 	cmd.Flags().StringVar(&jobID, "job-id", "", "ID of the job to download the debug bundle from")
 	cmd.Flags().BoolVar(&noConfirm, "no-confirm", false, "Disable confirmation prompt")
+	cmd.Flags().StringVar(&uploadURL, "upload-url", "", "If provided, where to upload the bundle in addition to creating a copy on disk")
 	return cmd
 }
